feat: Add ethlint support (#4)
* feat: Add ethlint support

* ci: Add contracts linter

* fix: Lint all contracts

+++ contracts/vaults/yWETH.sol
@@ -19,28 +19,30 @@ contract yWETH is ERC20, ERC20Detailed {
 
     IERC20 public token;
 
-    uint public min = 9990;
-    uint public constant max = 10000;
+    uint256 public min = 9990;
+    uint256 public constant max = 10000;
 
     address public governance;
     address public controller;
 
-    constructor (address _token, address _controller) public ERC20Detailed(
-        string(abi.encodePacked("yearn ", ERC20Detailed(_token).name())),
-        string(abi.encodePacked("y", ERC20Detailed(_token).symbol())),
-        ERC20Detailed(_token).decimals()
-    ) {
+    constructor(address _token, address _controller)
+        public
+        ERC20Detailed(
+            string(abi.encodePacked("yearn ", ERC20Detailed(_token).name())),
+            string(abi.encodePacked("y", ERC20Detailed(_token).symbol())),
+            ERC20Detailed(_token).decimals()
+        )
+    {
         token = IERC20(_token);
         governance = msg.sender;
         controller = _controller;
     }
 
-    function balance() public view returns (uint) {
-        return token.balanceOf(address(this))
-                .add(IController(controller).balanceOf(address(token)));
+    function balance() public view returns (uint256) {
+        return token.balanceOf(address(this)).add(IController(controller).balanceOf(address(token)));
     }
 
-    function setMin(uint _min) external {
+    function setMin(uint256 _min) external {
         require(msg.sender == governance, "!governance");
         min = _min;
     }
@@ -57,12 +59,12 @@ contract yWETH is ERC20, ERC20Detailed {
 
     // Custom logic in here for how much the vault allows to be borrowed
     // Sets minimum required on-hand to keep small withdrawals cheap
-    function available() public view returns (uint) {
+    function available() public view returns (uint256) {
         return token.balanceOf(address(this)).mul(min).div(max);
     }
 
     function earn() public {
-        uint _bal = available();
+        uint256 _bal = available();
         token.safeTransfer(controller, _bal);
         IController(controller).earn(address(token), _bal);
     }
@@ -71,13 +73,13 @@ contract yWETH is ERC20, ERC20Detailed {
         deposit(token.balanceOf(msg.sender));
     }
 
-    function deposit(uint _amount) public {
-        uint _pool = balance();
-        uint _before = token.balanceOf(address(this));
+    function deposit(uint256 _amount) public {
+        uint256 _pool = balance();
+        uint256 _before = token.balanceOf(address(this));
         token.safeTransferFrom(msg.sender, address(this), _amount);
-        uint _after = token.balanceOf(address(this));
+        uint256 _after = token.balanceOf(address(this));
         _amount = _after.sub(_before); // Additional check for deflationary tokens
-        uint shares = 0;
+        uint256 shares = 0;
         if (totalSupply() == 0) {
             shares = _amount;
         } else {
@@ -87,13 +89,13 @@ contract yWETH is ERC20, ERC20Detailed {
     }
 
     function depositETH() public payable {
-        uint _pool = balance();
-        uint _before = token.balanceOf(address(this));
-        uint _amount = msg.value;
+        uint256 _pool = balance();
+        uint256 _before = token.balanceOf(address(this));
+        uint256 _amount = msg.value;
         WETH(address(token)).deposit.value(_amount)();
-        uint _after = token.balanceOf(address(this));
+        uint256 _after = token.balanceOf(address(this));
         _amount = _after.sub(_before); // Additional check for deflationary tokens
-        uint shares = 0;
+        uint256 shares = 0;
         if (totalSupply() == 0) {
             shares = _amount;
         } else {
@@ -110,26 +112,25 @@ contract yWETH is ERC20, ERC20Detailed {
         withdrawETH(balanceOf(msg.sender));
     }
 
-
     // Used to swap any borrowed reserve over the debt limit to liquidate to 'token'
-    function harvest(address reserve, uint amount) external {
+    function harvest(address reserve, uint256 amount) external {
         require(msg.sender == controller, "!controller");
         require(reserve != address(token), "token");
         IERC20(reserve).safeTransfer(controller, amount);
     }
 
     // No rebalance implementation for lower fees and faster swaps
-    function withdraw(uint _shares) public {
-        uint r = (balance().mul(_shares)).div(totalSupply());
+    function withdraw(uint256 _shares) public {
+        uint256 r = (balance().mul(_shares)).div(totalSupply());
         _burn(msg.sender, _shares);
 
         // Check balance
-        uint b = token.balanceOf(address(this));
+        uint256 b = token.balanceOf(address(this));
         if (b < r) {
-            uint _withdraw = r.sub(b);
+            uint256 _withdraw = r.sub(b);
             IController(controller).withdraw(address(token), _withdraw);
-            uint _after = token.balanceOf(address(this));
-            uint _diff = _after.sub(b);
+            uint256 _after = token.balanceOf(address(this));
+            uint256 _diff = _after.sub(b);
             if (_diff < _withdraw) {
                 r = b.add(_diff);
             }
@@ -139,17 +140,17 @@ contract yWETH is ERC20, ERC20Detailed {
     }
 
     // No rebalance implementation for lower fees and faster swaps
-    function withdrawETH(uint _shares) public {
-        uint r = (balance().mul(_shares)).div(totalSupply());
+    function withdrawETH(uint256 _shares) public {
+        uint256 r = (balance().mul(_shares)).div(totalSupply());
         _burn(msg.sender, _shares);
 
         // Check balance
-        uint b = token.balanceOf(address(this));
+        uint256 b = token.balanceOf(address(this));
         if (b < r) {
-            uint _withdraw = r.sub(b);
+            uint256 _withdraw = r.sub(b);
             IController(controller).withdraw(address(token), _withdraw);
-            uint _after = token.balanceOf(address(this));
-            uint _diff = _after.sub(b);
+            uint256 _after = token.balanceOf(address(this));
+            uint256 _diff = _after.sub(b);
             if (_diff < _withdraw) {
                 r = b.add(_diff);
             }
@@ -159,11 +160,11 @@ contract yWETH is ERC20, ERC20Detailed {
         address(msg.sender).transfer(r);
     }
 
-    function getPricePerFullShare() public view returns (uint) {
+    function getPricePerFullShare() public view returns (uint256) {
         return balance().mul(1e18).div(totalSupply());
     }
 
-    function () external payable {
+    function() external payable {
         if (msg.sender != address(token)) {
             depositETH();
         }
