+++ interfaces/yearn/Strategy.sol
@@ -2,21 +2,20 @@
 
 pragma solidity ^0.5.16;
 
-
 interface Strategy {
     function want() external view returns (address);
 
     function deposit() external;
 
-    // NOTE: must exclude any tokens used in the yield 
+    // NOTE: must exclude any tokens used in the yield
     // Controller role - withdraw should return to Controller
     function withdraw(address) external;
 
     // Controller | Vault role - withdraw should always return to Vault
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 
     // Controller | Vault role - withdraw should always return to Vault
-    function withdrawAll() external returns (uint);
+    function withdrawAll() external returns (uint256);
 
-    function balanceOf() external view returns (uint);
+    function balanceOf() external view returns (uint256);
 }
