+++ interfaces/yearn/Token.sol
@@ -4,7 +4,9 @@ pragma solidity ^0.5.17;
 
 // NOTE: Basically an alias for Vaults
 interface yERC20 {
-  function deposit(uint256 _amount) external;
-  function withdraw(uint256 _amount) external;
-  function getPricePerFullShare() external view returns (uint);
+    function deposit(uint256 _amount) external;
+
+    function withdraw(uint256 _amount) external;
+
+    function getPricePerFullShare() external view returns (uint256);
 }
